+++ ShopWithMe/Models/SyncEntitaetsAlias.swift
@@ -37,6 +37,10 @@ final class SyncEntitaetsAlias {
 enum SyncEntitaetsArt {
     static let artikel = "Artikel"
     static let einkaufsliste = "Einkaufsliste"
+    static let geschaeft = "Geschaeft"
+    static let artikelKategorie = "ArtikelKategorie"
+    static let geschaeftTyp = "GeschaeftTyp"
+    static let kaufEintrag = "KaufEintrag"
 }
 
 enum SyncEntitaetsAliasService {